move Buf,BufMut implementations into a separate module
